move testdata out of engine, move e2e to execution/engine tests
more E2E than ever before!
