+++ v2/pkg/grpctest/schema.go
@@ -1,4 +1,4 @@
-package testdata
+package grpctest
 
 import (
 	"os"
